+++ libs/sdk-core/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Knowledge Box `title` and `slug` are required fields in `IKnowledgeBoxCreation` model
 - Adding optional `allowed_origins` field in `IKnowledgeBoxCreation` model
+- Improve reliability of resource creation and modification
 
 # 1.14.2 (2024-03-15)